+++ packages/rainbowkit/src/components/ConnectOptions/DesktopOptions.tsx
@@ -24,8 +24,10 @@ import { Text } from '../Text/Text';
 import {
   ConnectDetail,
   DownloadDetail,
+  DownloadOptionsDetail,
   GetDetail,
-  InstructionDetail,
+  InstructionExtensionDetail,
+  InstructionMobileDetail,
 } from './ConnectDetails';
 import {
   ScrollClassName,
@@ -38,8 +40,10 @@ export enum WalletStep {
   LearnCompact = 'LEARN_COMPACT',
   Get = 'GET',
   Connect = 'CONNECT',
+  DownloadOptions = 'DOWNLOAD_OPTIONS',
   Download = 'DOWNLOAD',
-  Instructions = 'INSTRUCTIONS',
+  InstructionsMobile = 'INSTRUCTIONS_MOBILE',
+  InstructionsExtension = 'INSTRUCTIONS_EXTENSION',
 }
 
 export function DesktopOptions({ onClose }: { onClose: () => void }) {
@@ -56,9 +60,9 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
   const compactModeEnabled = modalSize === ModalSizeOptions.COMPACT;
   const { disclaimer: Disclaimer } = useContext(AppContext);
 
-  const wallets = useWalletConnectors().filter(
-    wallet => wallet.ready || wallet.downloadUrls?.browserExtension
-  );
+  const wallets = useWalletConnectors()
+    .filter(wallet => wallet.ready || wallet.downloadUrls?.browserExtension)
+    .sort((a, b) => a.groupIndex - b.groupIndex);
 
   const groupedWallets = groupBy(wallets, wallet => wallet.groupName);
 
@@ -80,7 +84,7 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
     }
   };
 
-  const onSelectWallet = (wallet: WalletConnector) => {
+  const selectWallet = (wallet: WalletConnector) => {
     connectToWallet(wallet);
     setSelectedOptionId(wallet.id);
 
@@ -108,6 +112,23 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
           },
           uri ? 0 : 50
         );
+
+        // If the WalletConnect request is rejected, restart the wallet
+        // selection flow to create a new connection with a new QR code
+        const provider = await sWallet?.connector.getProvider();
+        const connection = provider?.signer?.connection;
+        if (connection?.on && connection?.off) {
+          const handleConnectionClose = () => {
+            removeHandlers();
+            selectWallet(wallet);
+          };
+          const removeHandlers = () => {
+            connection.off('close', handleConnectionClose);
+            connection.off('open', removeHandlers);
+          };
+          connection.on('close', handleConnectionClose);
+          connection.on('open', removeHandlers);
+        }
       });
     } else {
       setSelectedWallet(wallet);
@@ -115,11 +136,19 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
     }
   };
 
-  const getMobileWallet = (id: string) => {
+  const getWalletDownload = (id: string) => {
     setSelectedOptionId(id);
     const sWallet = wallets.find(w => id === w.id);
+    const isMobile = sWallet?.downloadUrls?.qrCode;
+    const isExtension = sWallet?.downloadUrls?.browserExtension;
     setSelectedWallet(sWallet);
-    changeWalletStep(WalletStep.Download);
+    if (isMobile && isExtension) {
+      changeWalletStep(WalletStep.DownloadOptions);
+    } else if (isMobile) {
+      changeWalletStep(WalletStep.Download);
+    } else {
+      changeWalletStep(WalletStep.InstructionsExtension);
+    }
   };
 
   const clearSelectedWallet = () => {
@@ -158,6 +187,11 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
     setConnectionError(false);
   }, [walletStep, selectedWallet]);
 
+  const hasExtensionAndMobile = !!(
+    selectedWallet?.downloadUrls?.browserExtension &&
+    (selectedWallet?.downloadUrls?.android || selectedWallet?.downloadUrls?.ios)
+  );
+
   switch (walletStep) {
     case WalletStep.None:
       walletContent = (
@@ -175,7 +209,7 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
       headerBackButtonLink = WalletStep.None;
       break;
     case WalletStep.Get:
-      walletContent = <GetDetail getMobileWallet={getMobileWallet} />;
+      walletContent = <GetDetail getWalletDownload={getWalletDownload} />;
       headerLabel = 'Get a Wallet';
       headerBackButtonLink = compactModeEnabled
         ? WalletStep.LearnCompact
@@ -204,6 +238,16 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
         ? clearSelectedWallet
         : () => {};
       break;
+    case WalletStep.DownloadOptions:
+      walletContent = selectedWallet && (
+        <DownloadOptionsDetail
+          changeWalletStep={changeWalletStep}
+          wallet={selectedWallet}
+        />
+      );
+      headerLabel = selectedWallet && `Get ${selectedWallet.name}`;
+      headerBackButtonLink = initialWalletStep;
+      break;
     case WalletStep.Download:
       walletContent = selectedWallet && (
         <DownloadDetail
@@ -211,24 +255,39 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
           wallet={selectedWallet}
         />
       );
-      headerLabel = hasQrCode && `Install ${selectedWallet.name}`;
-      headerBackButtonLink = initialWalletStep;
+      headerLabel = selectedWallet && `Install ${selectedWallet.name}`;
+      headerBackButtonLink = hasExtensionAndMobile
+        ? WalletStep.DownloadOptions
+        : initialWalletStep;
       break;
-    case WalletStep.Instructions:
+    case WalletStep.InstructionsMobile:
       walletContent = selectedWallet && (
-        <InstructionDetail
-          connectWallet={onSelectWallet}
+        <InstructionMobileDetail
+          connectWallet={selectWallet}
           wallet={selectedWallet}
         />
       );
       headerLabel =
-        hasQrCode &&
+        selectedWallet &&
         `Get started with ${
           compactModeEnabled
             ? selectedWallet.shortName || selectedWallet.name
             : selectedWallet.name
         }`;
       headerBackButtonLink = WalletStep.Download;
+      break;
+    case WalletStep.InstructionsExtension:
+      walletContent = selectedWallet && (
+        <InstructionExtensionDetail wallet={selectedWallet} />
+      );
+      headerLabel =
+        selectedWallet &&
+        `Get started with ${
+          compactModeEnabled
+            ? selectedWallet.shortName || selectedWallet.name
+            : selectedWallet.name
+        }`;
+      headerBackButtonLink = WalletStep.DownloadOptions;
       break;
     default:
       break;
@@ -304,8 +363,9 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
                             iconUrl={wallet.iconUrl}
                             key={wallet.id}
                             name={wallet.name}
-                            onClick={() => onSelectWallet(wallet)}
+                            onClick={() => selectWallet(wallet)}
                             ready={wallet.ready}
+                            recent={wallet.recent}
                           />
                         );
                       })}